move args to arguments.py; use self._precheck before perturbations; use _LOGGER instead of print
